refactor margin components to have a base layer
