+++ src/components/ui/margins/marginBase/index.ts
@@ -1,0 +1,1 @@
+export { Margin } from './margin';
