+++ src/components/ui/margins/marginLarge.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './marginBase';
 import { Spacing } from '../../../types/theme/Spacing';
 
 export const MarginLarge: React.FC = () => <Margin size={Spacing.Large} />;
