+++ src/components/ui/margins/marginMedium.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './marginBase';
 import { Spacing } from '../../../types/theme/Spacing';
 
 export const MarginMedium: React.FC = () => <Margin size={Spacing.Medium} />;
