+++ src/components/ui/margins/marginSmall.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './marginBase';
 import { Spacing } from '../../../types/theme/Spacing';
 
 export const MarginSmall: React.FC = () => <Margin size={Spacing.Small} />;
